requirements.txt
requests
pandas
beautifulsoup4
python-dotenv
spacy
fastapi
uvicorn
sqlalchemy
pydantic>=2.0
psycopg2-binary
passlib[bcrypt]
python-dotenv
pymysql
<<<<<<< HEAD
=======
pydantic[email]
>>>>>>> 683a7c8462bc2b233b657969544153859d880718


openai
scikit-learn
